requirements.txt
alabaster==0.7.13
arrow==1.3.0
asciitree==0.3.3
astroid==2.15.8
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.10.1
bleach==6.1.0
bokeh==3.1.1
cellpose==3.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.1
cookiecutter==2.4.0
cryptography==41.0.5
cupy-cuda12x==12.3.0
cycler==0.12.1
dask==2023.5.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
docstring-to-markdown==0.13
docutils==0.20.1
executing==2.0.1
fasteners==0.19
fastjsonschema==2.18.1
fastremap==1.15.0
fastrlock==0.8.2
ffmpeg-python==0.2.0
filelock==3.16.1
flake8==6.0.0
fonttools==4.43.1
fsspec==2023.10.0
future==1.0.0
h5py==3.10.0
idna==3.4
imagecodecs==2023.3.16
imageio==2.31.6
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.26.0
ipython==8.12.3
ipython-genutils==0.2.0
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
llvmlite==0.41.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.1
numcodecs==0.12.1
numpy==1.24.4
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.16.1
pylint==2.17.7
pylint-venv==3.0.3
pyls-spyder==0.4.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.7.4
python-slugify==8.0.1
pytoolconfig==1.2.6
pytz==2023.3.post1
PyWavelets==1.4.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.4
QtPy==2.4.1
rechunker==0.5.2
referencing==0.30.2
requests==2.31.0
rich==13.6.0
roifile==2023.5.12
roipoly==0.5.3
rope==1.10.0
rpds-py==0.10.6
Rtree==1.1.0
scikit-image==0.21.0
scipy==1.10.1
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.5
spyder-kernels==2.4.4
ssqueezepy @ git+https://github.com/OverLordGoldDragon/ssqueezepy@85563d024879d561ca4a56cd0c31a6e8d408cacb
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.0
three-merge==0.1.1
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.4.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==3.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2025.1
ujson==5.8.0
urllib3==2.0.7
watchdog==3.0.0
wcwidth==0.2.9
webencodings==0.5.1
whatthepatch==1.0.5
wrapt==1.15.0
wurlitzer==3.0.3
xyzservices==2025.1.0
yapf==0.40.2
zarr==2.16.1
zipp==3.17.0
